Store uploaded files in media folder

+++ bsyncviewer/tests/test_command.py
@@ -1,15 +1,16 @@
 import os
-from random import randint
-from shutil import copyfile
 from io import StringIO
+
+from django.conf import settings
+from django.core.files.uploadedfile import SimpleUploadedFile
 from django.core.management import call_command
 from django.test import TestCase
-from bsyncviewer.models.schema import Schema
+
 from bsyncviewer.models.bedes_models import BedesTerm, BedesEnumeration
+from bsyncviewer.models.schema import Schema
 from bsyncviewer.models.use_case import UseCase
 from bsyncviewer.models.use_case_attribute import UseCaseAttribute
 
-from django.conf import settings
 DEFAULT_SCHEMA_VERSION = settings.DEFAULT_SCHEMA_VERSION
 
 
@@ -28,59 +29,50 @@ class TestCommand(TestCase):
         for schema in schemas:
             schema.delete()
 
+
+class TestCommandWithSchema(TestCase):
+
+    def setUp(self):
+        self.schema = Schema.objects.filter(version=DEFAULT_SCHEMA_VERSION).first()
+        if not self.schema:
+            # add schema file - make sure to create a copy since the version will be deleted if
+            # the schema is deleted
+            sf = os.path.join(os.path.dirname(__file__), 'data', 'test_schema.xsd')
+            file = open(sf, 'rb')
+            simple_uploaded_file = SimpleUploadedFile(file.name, file.read())
+
+            self.schema = Schema(
+                name='Version {}'.format(DEFAULT_SCHEMA_VERSION),
+                version=DEFAULT_SCHEMA_VERSION,
+                schema_file=simple_uploaded_file
+            )
+            self.schema.save()  # Calling save also processes the schema and generates the template
+
     def test_create_use_case_command(self):
         print('TESTING CREATE USE CASE COMMAND')
-
-        # first add a schema
-        sf = os.path.join(os.path.dirname(__file__), 'data', 'test_schema.xsd')
-        schema_file = os.path.join(
-            os.path.dirname(__file__), 'data', 'schema_file_%s.xsd' % randint(0, 10000)
-        )
-        copyfile(sf, schema_file)
-
-        self.schema = Schema(
-            name='Version {}'.format(DEFAULT_SCHEMA_VERSION),
-            version=DEFAULT_SCHEMA_VERSION,
-            schema_file=schema_file
-        )
-        self.schema.save()  # Calling save also processes the schema and generates the template
 
         out = StringIO()
         call_command('create_use_case', stdout=out)
 
         # assert that a use case was created
-        useCases = UseCase.objects.all().count()
-        useCaseAttrs = UseCaseAttribute.objects.all().count()
-        print("USE CASES: {}, USE CASE ATTRIBUTES: {}".format(useCases, useCaseAttrs))
-        self.assertGreater(useCases, 0)
-        self.assertGreater(useCaseAttrs, 0)
-
-        # clean-up
-        self.schema.delete()
+        use_cases = UseCase.objects.all().count()
+        use_case_attrs = UseCaseAttribute.objects.all().count()
+        print("USE CASES: {}, USE CASE ATTRIBUTES: {}".format(use_cases, use_case_attrs))
+        self.assertGreater(use_cases, 0)
+        self.assertGreater(use_case_attrs, 0)
 
     def test_bedes_command(self):
+        # The schema must exist before this command is called.
         print('TESTING BEDES COMMAND')
-
-        # first add a schema
-        sf = os.path.join(os.path.dirname(__file__), 'data', 'test_schema.xsd')
-        schema_file = os.path.join(
-            os.path.dirname(__file__), 'data', 'schema_file_%s.xsd' % randint(0, 10000)
-        )
-        copyfile(sf, schema_file)
-
-        self.schema = Schema(
-            name='Version {}'.format(DEFAULT_SCHEMA_VERSION),
-            version=DEFAULT_SCHEMA_VERSION,
-            schema_file=schema_file
-        )
-        self.schema.save()  # Calling save also processes the schema and generates the template
 
         # create the CSV files
         out = StringIO()
-        call_command('bedes', schema_version=DEFAULT_SCHEMA_VERSION, bedes_version='v2.2', stdout=out)
+        call_command('bedes', schema_version=DEFAULT_SCHEMA_VERSION, bedes_version='v2.2',
+                     stdout=out)
 
         # add to database
-        call_command('bedes', schema_version=DEFAULT_SCHEMA_VERSION, bedes_version='v2.2', save_to_db=True, stdout=out)
+        call_command('bedes', schema_version=DEFAULT_SCHEMA_VERSION, bedes_version='v2.2',
+                     save_to_db=True, stdout=out)
 
         # check that there are items in bedes models
         bterms = BedesTerm.objects.all().count()
@@ -89,5 +81,7 @@ class TestCommand(TestCase):
         self.assertGreater(bterms, 0)
         self.assertGreater(benums, 0)
 
-        # clean-up
-        self.schema.delete()
+    def tearDown(self):
+        # clean-up files on disk
+        if self.schema and self.schema.id is not None:
+            self.schema.delete()
